Keep `TastyVersion` in `CompilationUnitInfo`
